Refactor out task queue
