+++ src/instance/app.cfg.example
@@ -98,12 +98,9 @@ ENTITY_WEBSERVICE_URL = 'https://entity-api.dev.hubmapconsortium.org'
 # URL to the UBKG web service (shown for DEV)
 UBKG_WEBSERVICE_URL = 'https://ontology-api.dev.hubmapconsortium.org/'
 
-# URL to the Redis server used by the cell_count_type computation queuing mechanism.
-# Use port 6379 exposed on the ingest-api-rq-server container, not the port on host
-REDIS_URL = 'redis://ingest-api-rq-server:6379'
-# For local debugging....
-# $ docker run -d -p 6379:6379 --name rq-server redis
-# REDIS_URL = 'redis://localhost:6379'
+# URL to the Redis server
+# Use port 6379 exposed on the ingest-api-redis-server container, not the port on host
+REDIS_URL = 'redis://ingest-api-redis-server:6379'
 
 # For UI login/logout
 # Flask app session key
