+++ src/requirements.txt
@@ -21,13 +21,8 @@ requests==2.32.3
 # git+https://github.com/hubmapconsortium/commons.git@${COMMONS_BRANCH}#egg=hubmap-commons
 hubmap-commons==2.1.19
 
-# For parsing the secondary analysis files...
-anndata==0.7.8
-
-# For queuing of extract cell type count tasks...
+# For /datasets/data-status and /uploads/data-status cache
 redis==4.3.4
-rq==1.11.1
-rq-dashboard==0.6.1
 
 # For assay type rules
 rule_engine==4.1.0
